+++ src/schema/algolia.js
@@ -5,7 +5,36 @@ import resolvers from '../resolvers/algolia';
 
 const typeDefs = gql`
   type Query {
-    searchCampaigns: [Int]!
+    # searchCampaigns: [Int]!
+
+    searchCampaigns(
+      term: String
+      isOpen: Boolean
+      perPage: Int
+      cursor: String
+    ): AlgoliaCollection!
+
+    # searchPages: [Int]!
+  }
+
+  type AlgoliaCollection {
+    results: [Int]!
+  }
+
+  # type AlgoliaCollection {
+  #   edges: [SearchEdge]
+  #   pageInfo: PageInfo!
+  # }
+
+  type SearchEdge {
+    cursor: String!
+    node: Int!
+  }
+
+  type PageInfo {
+    endCursor: String
+    hasNextPage: Boolean!
+    hasPreviousPage: Boolean!
   }
 `;
 
@@ -15,7 +44,7 @@ export default makeExecutableSchema({
 });
 
 // {
-//   hits: [],
+//   hits: [...],
 //   nbHits: 183,
 //   page: 0,
 //   nbPages: 10,
